@@ -38,38 +38,61 @@
   "author": "Lee Driscoll <lee.driscoll@orbistech.co.uk>",
   "license": "ISC",
   "dependencies": {
-    "ava": "^0.19.0",
+    "ava": "^0.19.1",
+    "babel-cli": "^6.24.1",
+    "babel-eslint": "^7.2.3",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-polyfill": "^6.23.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
     "bunyan": "^1.8.8",
+    "core-decorators": "^0.18.0",
     "cross-env": "^3.2.3",
+    "debug": "^2.6.6",
+    "grommet": "^1.4.0",
+    "grommet-addons": "^0.3.0",
     "history": "^4.6.1",
     "isomorphic-fetch": "^2.2.1",
     "localforage": "^1.5.0",
     "localforage-cordovasqlitedriver": "^1.6.0",
     "lodash": "^4.17.4",
+    "nock": "^9.0.13",
     "pouch-redux-middleware": "^0.6.1",
-    "pouchdb": "^6.1.2",
+    "pouchdb": "^6.2.0",
+    "progress-promise": "0.0.6",
+    "prop-types": "^15.5.9",
     "quagga": "^0.11.6",
-    "react": "^15.4.2",
+    "react": "^15.5.4",
     "react-atellier": "^0.1.13",
-    "react-dom": "^15.4.2",
-    "react-redux": "^5.0.3",
-    "react-router": "^3.0.0",
-    "react-router-redux": "^5.0.0-alpha.4",
+    "react-dom": "^15.5.4",
+    "react-redux": "^5.0.4",
+    "react-router": "^3.0.5",
+    "react-router-dom": "^4.1.1",
+    "react-router-redux": "^5.0.0-alpha.6",
     "redux": "^3.6.0",
+    "redux-devtools": "^3.4.0",
+    "redux-devtools-dock-monitor": "^1.1.2",
+    "redux-devtools-extension": "^2.13.2",
+    "redux-devtools-log-monitor": "^1.3.0",
     "redux-logger": "^3.0.1",
-    "redux-persist": "^4.6.0",
+    "redux-persist": "^4.8.0",
     "redux-persist-transform-filter": "0.0.9",
-    "redux-thunk": "^2.2.0"
+    "redux-thunk": "^2.2.0",
+    "sass-loader": "^6.0.5",
+    "webpack": "^2.5.1",
+    "webpack-dev-middleware": "^1.10.2"
   },
   "devDependencies": {
     "ava": "^0.18.2",
     "babel-cli": "^6.24.0",
     "babel-eslint": "^7.2.1",
     "babel-plugin-add-module-exports": "^0.2.1",
+    "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-object-rest-spread": "^6.23.0",
     "babel-preset-es2015": "^6.24.0",
     "babel-preset-react": "^6.23.0",
+    "babel-preset-stage-0": "^6.24.1",
     "body-parser": "^1.17.1",
     "cordova": "^6.5.0",
     "css-loader": "^0.27.3",
@@ -103,7 +126,7 @@
     "uuid": "^3.0.1",
     "webpack": "^2.2.1",
     "webpack-dashboard": "^0.3.0",
-    "webpack-dev-middleware": "^1.10.1",
+    "webpack-dev-middleware": "^1.10.2",
     "webpack-hot-middleware": "^2.17.1",
     "xo": "^0.17.1"
   },
