@@ -1,5 +1,6 @@
 import fetch from 'isomorphic-fetch'
-import { v4 } from 'uuid'
+import { v4 as uuidGen } from 'uuid'
+import { find, filter, includes } from 'lodash'
 import { apiURL } from '../config'
 import {
   ADD_ORDER, DELETE_ORDER, CHANGE_ORDER_QUANTITY,
@@ -8,6 +9,7 @@ import {
   CHANGE_OPERATION_MODE
 } from '../constants/ActionTypes'
 import { checkStatusAndParseJSON } from '../helpers/Network'
+import { failIfMissing } from '../helpers/Function'
 
 export function addOrder(id, order) {
   return { type: ADD_ORDER, id, order }
@@ -40,34 +42,38 @@ export function succeedProcessOrders() {
   }
 }
 
-export function failProcessOrders() {
+export function failProcessOrders(error) {
   return {
-    type: FAIL_PROCESS_ORDERS
+    type: FAIL_PROCESS_ORDERS,
+    error
   }
 }
 
-export function processOrders(sessionID, orders, device) {
-  return function (dispatch) {
+export function processOrders() {
+  return function (dispatch, getState) {
     dispatch(requestProcessOrders())
+
+    const orders = getState().orderEntities
 
     return fetch(apiURL, {
       method: 'post',
-      body: {
+      headers: {'Content-Type':'application/x-www-form-urlencoded'},
+      body: JSON.stringify({
         method: 'HandheldService.ProcessTransactions',
         params: {
-          SessionID: sessionID,
-          Data: orders
+          SessionID: getState().session.id,
+          Data: getState().orders.unprocessedItems.map(id => {
+            return orders.hasOwnProperty(id) && orders[id]
+          })
         }
-      }
+      })
     })
       .then(checkStatusAndParseJSON)
       .then(() => {
-        dispatch(receiveProcessOrders())
         dispatch(succeedProcessOrders())
+        dispatch(receiveProcessOrders())
       })
-      .catch(error => {
-        dispatch(failProcessOrders(error))
-      })
+      .catch(error => dispatch(failProcessOrders(error)))
   }
 }
 
@@ -79,22 +85,42 @@ export function changeOperationMode(mode) {
 }
 
 export function createTransaction(mode, barcode, quantity) {
-  return function (dispatch) {
+  return function (dispatch, getState) {
+    const transaction = findTransactionByBarcode(getState, barcode, mode)
 
-    const order = {
-      __type: 'HandheldTrans',
-      AreaID: '',
-      Barcode: barcode,
-      Qty: quantity,
-      Ref1: '',
-      Ref2: '',
-      // TermianlID: store.getState().terminalID,
-      // TransDate: now.toISOString().substr(-1),
-      TransType: mode,
-      UnitID: '',
-      // UserID: store.getState().user.id
+    if (transaction) {
+      dispatch(changeOrderQuantity(transaction._id, transaction.Qty + quantity))
+    } else {
+      const uuid = uuidGen()
+
+      const order = {
+        _id: uuid,
+        __type: 'HandheldTrans',
+        AreaID: '',
+        Barcode: barcode,
+        Qty: quantity,
+        Ref1: '',
+        Ref2: '',
+        TermianlID: getState().terminalID,
+        TransDate: new Date().toISOString().slice(0, -1),
+        TransType: mode,
+        UnitID: '',
+        UserID: getState().user.id
+      }
+
+      dispatch(addOrder(uuid, order))
     }
-
-    dispatch(addOrder(uuidGen(), order))
   }
+}
+
+function findTransactionByBarcode(getState, barcode, mode) {
+  return find(
+    filter(getState().orderEntities, (item, key) => {
+      return includes(getState().orders.unprocessedItems, key)
+    }),
+    {
+      TransType: mode,
+      Barcode: barcode
+    }
+  )
 }
