@@ -1,13 +1,11 @@
-import React, { Component, PropTypes } from 'react'
+import React, { Component } from 'react'
+import PropTypes from 'prop-types'
 import Menu from 'grommet/components/Menu'
-import Button from 'grommet/components/Button'
 import Anchor from 'grommet/components/Anchor'
 import Header from 'grommet/components/Header'
-import Heading from 'grommet/components/Heading'
 import Modes from '../constants/OperationModes'
 
 class ModeSwitcher extends Component {
-
   constructor(props) {
     super(props)
 
@@ -26,31 +24,23 @@ class ModeSwitcher extends Component {
 
   render() {
     return (
-      <Header
-        direction='row'
-        justify='center'>
-        <Menu
-          label='Change Mode'
-          inline={false}
-          direction='row'
-          justify='between'
-          size='small'
-          flex='grow'
-          responsive={false}>
-          <Anchor
-            label='Stocktake'
-            onClick={this.onSwitch.bind(this, Modes.STOCKTAKE)}
-            className={this.state.mode === Modes.STOCKTAKE && 'grommetux-anchor--primary'}/>
-          <Anchor
-            label='Order'
-            onClick={this.onSwitch.bind(this, Modes.ORDERING)}
-            className={this.state.mode === Modes.ORDERING && 'grommetux-anchor--primary'}/>
-          <Anchor
-            label='Delivery'
-            onClick={this.onSwitch.bind(this, Modes.DELIVERY)}
-            className={this.state.mode === Modes.DELIVERY && 'grommetux-anchor--primary'}/>
-        </Menu>
-      </Header>
+      <Menu primary={true}>
+        <Anchor
+          label='Stocktake'
+          onClick={this.onSwitch.bind(this, Modes.STOCKTAKE)}
+          className={this.state.mode === Modes.STOCKTAKE && 'active'}
+          />
+        <Anchor
+          label='Order'
+          onClick={this.onSwitch.bind(this, Modes.ORDERING)}
+          className={this.state.mode === Modes.ORDERING && 'active'}
+          />
+        <Anchor
+          label='Delivery'
+          onClick={this.onSwitch.bind(this, Modes.DELIVERY)}
+          className={this.state.mode === Modes.DELIVERY && 'active'}
+          />
+      </Menu>
     )
   }
 }
