@@ -1,26 +1,22 @@
 import React, { Component } from 'react'
-import { Router, Route } from 'react-router'
+import { Router, Route, IndexRoute } from 'react-router'
 import history from '../history'
 import LoginContainer from '../containers/LoginContainer.jsx'
 import AppLayout from '../components/AppLayout.jsx'
-import MainLayout from '../components/MainLayout.jsx'
-import OrdersLayout from '../components/OrdersLayout.jsx'
+import OrdersLayoutContainer from '../containers/OrdersLayoutContainer'
+import styles from '../assets/scss/orbis/index.scss'
 
 const routes = (
   <Route component={AppLayout}>
     <Route exact path="/" component={LoginContainer}/>
-    <Route component={MainLayout}>
-      <Route path="/orders" component={OrdersLayout}/>
-    </Route>
+    <Route path="/orders" component={OrdersLayoutContainer}/>
   </Route>
 )
 
 export default class Root extends Component {
-
   render() {
     return (
       <Router history={history} routes={routes}/>
     )
   }
-
 }
