@@ -8,8 +8,8 @@ const mapStateToProps = state => {
       return state.orderEntities[id] &&
         state.orderEntities[id].TransType === state.orders.mode
     }).map(id => {
-      const barcode = state.barcodeEntities[id]
-      let order = state.orderEntities[id]
+      const order = state.orderEntities[id]
+      const barcode = state.barcodeEntities[order.Barcode]
 
       if (order) {
         if (barcode) {
