@@ -2,9 +2,11 @@ import { orders, orderEntities } from './reducers/OrdersReducer'
 import { products, productEntities } from './reducers/ProductsReducer'
 import { cashiers, cashierEntities } from './reducers/CashiersReducer'
 import { barcodes, barcodeEntities } from './reducers/BarcodesReducer'
+import barcodeLookup from './reducers/BarcodeLookupReducer'
 import { session, user } from './reducers/SessionReducer'
 import sync from './reducers/SyncReducer'
 import scanner from './reducers/ScannerReducer'
+import ui from './reducers/UIReducer'
 
 export {
   orders,
@@ -15,8 +17,10 @@ export {
   cashierEntities,
   barcodes,
   barcodeEntities,
+  barcodeLookup,
   session,
   user,
   scanner,
-  sync
+  sync,
+  ui
 }
