@@ -1,5 +1,6 @@
 export default function sync(state = {
-  isSyncing: false
+  isSyncing: false,
+  progress: 0
 }, action) {
   switch (action.type) {
     case 'START_SYNC':
@@ -11,6 +12,11 @@ export default function sync(state = {
       return {
         ...state,
         isSyncing: false
+      }
+    case 'SYNC_PROGRESS':
+      return {
+        ...state,
+        progress: action.progress
       }
     default:
       return state
